Show tournament entry screenshots on party page
